pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "artifacta"
version = "0.1.0"
description = "Universal experiment and artifact tracking — gain insights and optimize models with confidence"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Elastic License 2.0"}
authors = [
    {name = "Artifacta Contributors"}
]
keywords = ["experiment-tracking", "visualization", "mlops", "data-science"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.0.0",
    "flask>=2.0.0",
    "flask-cors>=3.0.0",
    "pillow>=9.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "pydocstyle>=6.3.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-autodoc-typehints>=1.24.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.0",
    "pillow>=9.0.0",
    "pytorch-lightning>=2.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tensorflow>=2.13.0",
    "reportlab>=4.0.0",
    "av>=10.0.0",
    "bump-my-version>=0.26.0",
]

[project.scripts]
artifacta = "tracking_server.cli:cli"

[project.urls]
Homepage = "https://github.com/walkerbdev/artifacta"
Documentation = "https://github.com/walkerbdev/artifacta#readme"
Repository = "https://github.com/walkerbdev/artifacta"
Issues = "https://github.com/walkerbdev/artifacta/issues"

[tool.setuptools]
packages = ["artifacta", "tracking_server", "artifacta_ui"]
package-dir = {"artifacta" = "artifacta/artifacta", "tracking_server" = "tracking-server"}
include-package-data = true

[tool.setuptools.package-data]
artifacta_ui = ["dist/**/*", "index.html"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

# Mypy configuration
[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
show_error_codes = true

# Ignore missing imports for third-party libraries without type stubs
[[tool.mypy.overrides]]
module = ["pydantic.*", "click.*", "fastapi.*", "uvicorn.*"]
ignore_missing_imports = true
disallow_untyped_decorators = false

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers"

# Bump My Version configuration
[tool.bumpversion]
current_version = "0.1.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
commit_args = ""

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "docs/conf.py"
search = 'release = "{current_version}"'
replace = 'release = "{new_version}"'

[[tool.bumpversion.files]]
filename = "tracking-server/cli.py"
search = 'version="{current_version}"'
replace = 'version="{new_version}"'
